pyproject.toml
[project]
name = "bankapi"
version = "0.1.0"
description = ""
authors = [
    {name = "cezarcorrea",email = "cezar.correa1234@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "databases (>=0.9.0,<0.10.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "tzlocal (>=5.3.1,<6.0.0)",
    "slowapi (>=0.1.9,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
